# openapispec.yaml
openapi: 3.1.0
info:
  title: David Chang API
  summary: Programmatic way of interacting with me!
  description: Programmatic way of interacting with me!
  version: 0.0.3
  contact:
    name: David Chang
    url: 'https://davidchang.dev'
    email: email@davidchang.dev
  termsOfService: 'https://davidchang.dev/'
  license:
    name: The MIT License
    url: 'https://opensource.org/license/mit/'
servers:
  - url: 'https://api.davidchang.dev'
    description: Production
paths:
  /resumes:
    get:
      operationId: getResumes
      summary: Get all resumés
      description: Simple API that returns a JSON array of all available resumés
      tags:
        - resumes
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumesResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: addResume
      summary: Add new resumé
      description: Simple API to add a new resumé to the database
      tags:
        - resumes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/resumes/{id}':
    get:
      operationId: getResumeById
      summary: Get resumé by ID
      description: Simple API to get a single resumé by ID
      tags:
        - resumes
      parameters:
        - name: id
          description: Autoincremented value uniquely identifying the resumé
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeResponse'
        '404':
          description: Resumé not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: updateResumeByID
      summary: Update resumé by ID
      description: Simple API to update an existing resumé by ID
      tags:
        - resumes
      parameters:
        - name: id
          description: Autoincremented value uniquely identifying the resumé
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resumé not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/resumes/latest':
    get:
      operationId: getLatestResume
      summary: Get the most recent resumé
      description: API to get the most recent resumé by year. In case of a tie, returns the most recently modified one.
      tags:
        - resumes
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeResponse'
        '404':
          description: No resumes found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Resume:
      type: object
      properties:
        id:
          type: string
        year:
          type: integer
        URL:
          type: string
    ResumeRequest:
      type: object
      properties:
        year:
          type: integer
        URL:
          type: string
    ResumeResponse:
      allOf:
        - $ref: '#/components/schemas/Resume'
        - type: object
          properties:
            id:
              type: string
    ResumesResponse:
      type: array
      items:
        $ref: '#/components/schemas/Resume'
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
tags:
  - name: documentation
    description: Documentation for api.davidchang.dev
  - name: resumes
    description: Resumé data